Рефакторинг 2 notification.py
