+++ app/tests/startTests.sh
@@ -1,4 +1,4 @@
 #! /usr/bin/env bash
-echo "Waiting 3 seconds and run test"
-sleep 3
-pytest  --disable-warnings -v
+echo "Waiting 1 seconds and run test"
+sleep 1
+pytest -v -s --cache-clear --log-level=INFO
